requirements.txt
torch>=1.9.0
numpy>=1.19.5
pandas>=1.3.0
matplotlib>=3.4.3
seaborn>=0.11.2
jupyter>=1.0.0
gym==0.19.0
gym-idsgame>=1.0.0
scikit-learn>=0.24.2
pytest>=6.2.5
black>=21.7b0
flake8>=3.9.2
isort>=5.9.3
python-dotenv>=0.19.0
tensorboard>=2.6.0
tqdm>=4.62.2